Refactoring the TranslateBehavior to use engines.
